adverscan/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
numpy>=1.24.0
pydantic>=2.5.0
scikit-learn>=1.3.0
joblib>=1.3.2
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
xgboost>=2.0.0

